refactor: utils extension
